requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
google-generativeai
python-dotenv
Werkzeug
reportlab
Pillow